OrchidOs : initial HTML structure with metadata, CSS links, Font Awesome icons, Animate CDN and jQuery script.
Created a boilerplate template for website development. The structure includes folders for videos and images, along with placeholder files. HTML files are prefilled with CDN links and JavaScript imports, including Bootstrap and Animate.css.  Also included a basic page load log for the debug branch to ensure proper loading. that will show in the web browser console on loading the page
Note : more information can be found in the extra-info folder .